tests: Move wrongly placed tests to the correct class.
